+++ types/types.ts
@@ -9,6 +9,7 @@ export type LucraEnvironment =
 export enum LucraClientMessageType {
   achievementsResponse = "achievementsResponse",
   activeMatchupStarted = "activeMatchupStarted",
+  autoJoinedTournaments = "autoJoinedTournaments",
   claimReward = "claimReward",
   convertToCredit = "convertToCredit",
   deepLink = "deepLink",
@@ -37,6 +38,7 @@ export enum LucraClientMessageType {
 
 export enum MessageTypeToLucraClient {
   achievementsRequest = "achievementsRequest",
+  autoJoinTournamentsRequest = "autoJoinTournamentsRequest",
   availableRewards = "availableRewards",
   clientUserInfo = "clientUserInfo",
   convertToCreditResponse = "convertToCreditResponse",
@@ -82,6 +84,7 @@ export type LucraMatchupCreatedBody = { matchupId: string };
 export type LucraMatchupCanceledBody = { matchupId: string };
 export type LucraMatchupAcceptedBody = { matchupId: string };
 export type LucraTournamentJoinedBody = { matchupId: string };
+export type LucraAutoJoinedTournamentsBody = { tournamentIds: string[] };
 export type LucraConvertToCreditBody = { amount: number };
 export type LucraDeepLinkBody = { url: string };
 export type LucraMatchupInviteUrlBody = { matchupId: string };
@@ -122,6 +125,7 @@ export type LucraClientConstructor = {
   tenantId: string;
   env: LucraEnvironment;
   locationId?: string;
+  autoJoin?: boolean;
 };
 
 export type LucraEventMap = {
@@ -133,6 +137,7 @@ export type LucraEventMap = {
   matchupCanceled: LucraMatchupCanceledBody;
   convertToCredit: LucraConvertToCreditBody;
   tournamentJoined: LucraTournamentJoinedBody;
+  autoJoinedTournaments: LucraAutoJoinedTournamentsBody;
   deepLink: LucraDeepLinkBody;
   navigationEvent: LucraNavigationEventBody;
   claimReward: LucraClaimRewardBody;